Added ltree to comments, added js to prevent hotkeys, redone private controller
