+++ Gemfile
@@ -16,6 +16,7 @@ gem 'http_accept_language'
 gem 'jbuilder'
 gem 'jquery-rails'
 gem 'jquery-ui-rails'
+gem 'pg_ltree', '1.1.7'
 gem 'puma'
 gem 'rails', '~> 5.2'
 gem 'rails-i18n'
