+++ app/controllers/private/domains_controller.rb
@@ -20,10 +20,6 @@ class Private::DomainsController < PrivateController
 
   private
 
-  def set_owned_project
-    @project = current_user.owned_projects.find(params[:project_id])
-  end
-
   def set_project_domain_service
     @project_domain_service = ProjectDomainService.new(area_private_domain: area_private_domain, project: @project)
   end
