+++ app/controllers/private/projects_controller.rb
@@ -11,6 +11,7 @@ class Private::ProjectsController < PrivateController
   end
 
   def show
+    byebug
     page = @project.pages.first
 
     redirect_to private_project_page_path(@project, page)
@@ -71,13 +72,5 @@ class Private::ProjectsController < PrivateController
       :google_analytics_tracker_id,
       :private, :secret_enabled
     )
-  end
-
-  def set_project
-    @project = current_user.projects.find(params[:id])
-  end
-
-  def set_owned_project
-    @project = current_user.owned_projects.find(params[:id])
   end
 end
