+++ app/controllers/private_controller.rb
@@ -6,4 +6,16 @@ class PrivateController < ApplicationController
   def public_controller?
     false
   end
+
+  def set_page
+    @page = @project.pages.find(params[:id])
+  end
+
+  def set_project
+    @project = current_user.projects.find(params[:project_id])
+  end
+
+  def set_owned_project
+    @project = current_user.owned_projects.find(params[:project_id])
+  end
 end
