+++ app/views/private/comments/_comment.html.erb
@@ -1,4 +1,5 @@
-<div class="px-4 mt-3 mb-5">
+<% comment_padding = comment.depth >= 9 ? 'mt-3 mb-5' : 'pl-4 mt-3 mb-5' %>
+<div class="<%= comment_padding %>">
   <% if comment.deleted? %>
     <strong><%= t('comments.user_deleted') %></strong>
     <div class="small text-muted mb-2">
@@ -15,23 +16,26 @@
   <% end %>
 
   <% if signed_in? %>
-    <div class="mt-n2">
-      <%= link_to t('comments.answer'), "#commentCollapse_#{dom_id(comment)}",
-                  data: { toggle: :collapse },
-                  aria: { expanded: false, controls: "commentCollapse_#{dom_id(comment)}" } %>
+    <% unless comment.deleted? %>
+      <div class="mt-n2">
+        <%= link_to t('comments.answer'), "#commentCollapse_#{dom_id(comment)}",
+                    data: { toggle: :collapse },
+                    aria: { expanded: false, controls: "commentCollapse_#{dom_id(comment)}" } %>
 
-      <%= link_to t('comments.delete'), [:private, @project, @page, comment],
-                  class: "ml-2",
-                  method: :delete,
-                  data: { confirm: t('comments.delete_confirmation') } if comment.user == current_user %>
-
-      <%= render partial: "private/comments/form", locals: { parent_id: comment.id,
-                                                             id: "commentCollapse_#{dom_id(comment)}",
-                                                             class: "collapse" } %>
+        <%= link_to t('comments.delete'), [:private, @project, @page, comment],
+                    class: "ml-2",
+                    method: :delete,
+                    data: { confirm: t('comments.delete_confirmation') } if comment.user == current_user %>
+      </div>
+    <% end %>
+    <div>
+    <%= render partial: "private/comments/form", locals: { parent_id: comment.id,
+                                                           id: "commentCollapse_#{dom_id(comment)}",
+                                                           class: "collapse" } %>
     </div>
   <% end %>
 
   <%= tag.div id: "#{dom_id(comment)}_comments" do %>
-    <%= render partial: "private/comments/comment", collection: comment.comments %>
+    <%= render partial: "private/comments/comment", collection: comment.children %>
   <% end %>
 </div>
